api/controller/requirements.txt
asgiref==3.7.2
asttokens==2.4.1
backcall==0.2.0
backports.zoneinfo==0.2.1
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.2.0
debugpy==1.8.0
decorator==5.1.1
Django==4.2.8
django-cors-headers==4.3.1
djangorestframework==3.14.0
executing==2.0.1
idna==3.6
importlib-metadata==7.0.1
ipykernel==6.28.0
ipython==8.12.3
jedi==0.19.1
jupyter_client==8.6.0
jupyter_core==5.5.1
matplotlib-inline==0.1.6
nest-asyncio==1.5.8
packaging==23.2
parso==0.8.3
pickleshare==0.7.5
platformdirs==4.1.0
prompt-toolkit==3.0.43
psutil==5.9.7
pure-eval==0.2.2
Pygments==2.17.2
python-dateutil==2.8.2
pytz==2023.3.post1
pywin32==306
pyzmq==25.1.2
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
tornado==6.4
traitlets==5.14.0
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
wcwidth==0.2.12
zipp==3.17.0
